Split file parsing/encryption/verification into separate files.
