+++ Sming/Libraries/OtaUpgrade/appcode/OtaUpgradeData.cpp
@@ -1,5 +1,14 @@
 #include <FlashString/Array.hpp>
 
+DEFINE_FSTR_ARRAY(OTAUpgrade_AppFlashRegionOffsets, uint32_t,
+#ifdef RBOOT_SPIFFS_0
+				  RBOOT_SPIFFS_0,
+#endif
+#ifdef RBOOT_SPIFFS_1
+				  RBOOT_SPIFFS_1,
+#endif
+				  0)
+
 #ifdef ENABLE_OTA_ENCRYPTION
 IMPORT_FSTR_ARRAY(OTAUpgrade_DecryptionKey, uint8_t, PROJECT_DIR "/out/OtaUpgrade/decrypt.key.bin");
 #endif
@@ -7,13 +16,3 @@ IMPORT_FSTR_ARRAY(OTAUpgrade_DecryptionKey, uint8_t, PROJECT_DIR "/out/OtaUpgrad
 #ifdef ENABLE_OTA_SIGNING
 IMPORT_FSTR_ARRAY(OTAUpgrade_SignatureVerificationKey, uint8_t, PROJECT_DIR "/out/OtaUpgrade/verify.key.bin");
 #endif
-
-DEFINE_FSTR_ARRAY(OTAUpgrade_AppFlashRegionOffsets, uint32_t,
-#ifdef RBOOT_SPIFFS_0
-    RBOOT_SPIFFS_0,
-#endif
-#ifdef RBOOT_SPIFFS_1
-    RBOOT_SPIFFS_1,
-#endif
-    0
-)
